re-work of calibration done.. but still time is not accurate
